+++ pkg/commands/logging/gcs/update.go
@@ -116,7 +116,7 @@ func (c *UpdateCommand) ConstructInput(serviceID string, serviceVersion int) (*f
 		input.CompressionCodec = &c.CompressionCodec.Value
 	}
 	if c.Format.WasSet {
-		input.Format = &c.Format.Value
+		input.Format = fastly.ToPointer(argparser.Content(c.Format.Value))
 	}
 	if c.FormatVersion.WasSet {
 		input.FormatVersion = &c.FormatVersion.Value