+++ pkg/commands/logging/googlepubsub/create.go
@@ -98,7 +98,7 @@ func (c *CreateCommand) ConstructInput(serviceID string, serviceVersion int) (*f
 		input.Name = &c.EndpointName.Value
 	}
 	if c.Format.WasSet {
-		input.Format = &c.Format.Value
+		input.Format = fastly.ToPointer(argparser.Content(c.Format.Value))
 	}
 	if c.FormatVersion.WasSet {
 		input.FormatVersion = &c.FormatVersion.Value