+++ pkg/commands/logging/heroku/create.go
@@ -95,7 +95,7 @@ func (c *CreateCommand) ConstructInput(serviceID string, serviceVersion int) (*f
 	}
 
 	if c.Format.WasSet {
-		input.Format = &c.Format.Value
+		input.Format = fastly.ToPointer(argparser.Content(c.Format.Value))
 	}
 
 	if c.FormatVersion.WasSet {