+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - feat(config): Support file/format for kv_store and secret_store in fastly.toml
 - feat(config): Support metadata for kv_store in fastly.toml
+- feat(logging): add support for passing a filepath as an argument for format in logging commands
 
 ### Bug fixes: